+++ BraintreeDropIn/Custom Views/BTEnrollmentVerificationViewController.m
@@ -32,6 +32,9 @@
     self.title = BTUIKLocalizedString(CONFIRM_ENROLLMENT_LABEL);
     
     self.view.backgroundColor = [BTUIKAppearance sharedInstance].formBackgroundColor;
+    if (@available(iOS 15.0, *)) {
+        self.navigationController.navigationBar.scrollEdgeAppearance = self.navigationController.navigationBar.standardAppearance;
+    }
     self.navigationController.navigationBar.barTintColor = [BTUIKAppearance sharedInstance].barBackgroundColor;
     self.navigationController.navigationBar.translucent = NO;
     [self.navigationController.navigationBar setTitleTextAttributes:@{